requirements.txt
fastapi
streamlit==1.45.1
yfinance
requests
beautifulsoup4
pandas
numpy==1.26.4
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.29.2
huggingface-hub==0.16.4
langchain==0.1.14
langgraph==0.0.40
crewai==0.11.2
python-dotenv
uvicorn
pytest
pyttsx3
regex==2023.12.25
packaging==23.2
st-audiorec
groq==0.9.0